PS4 port: add PS4 conditionals and convert cmake code generation to add_custom_target
The add_custom_command() functions in cmake were not working for the PS4
cmake project, so the generated cpp files were not being created. Switched
over to a add_custom_target, which worked. Added PS4 to a bunch of cmake
conditionals already used for various platforms, and __ORBIS__ in C++ files
for specific compiler conditionals for fancy C++ features.